+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # [0.0.70] - In progress
 * update Builder3 logger
-* create markdown generator on Builder3
+* update markdown generator on Builder3
 * 
 # [0.0.69] - 2024-05-06
 * add toc generator for html and refactoring, Builder3