+++ changelog.md
@@ -1,13 +1,14 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+# [0.0.71] - In progress
 
-# [0.0.70] - In progress
+# [0.0.70] - 2024-05-13
 * update Builder3 logger
 * update markdown generator on Builder3
 * pushing files to FileGroup optimization (Builder3)
 * optimize aggregated file group in Builder3
-* 
+  
 # [0.0.69] - 2024-05-06
 * add toc generator for html and refactoring, Builder3
 * add static markdown generation, Builder3